requirements.txt
# FastAPI & server
fastapi==0.115.13
uvicorn[standard]
psutil
python-multipart
pydantic
requests

# HuggingFace model
transformers==4.52.4
huggingface-hub==0.33.0
safetensors==0.5.3
timm==1.0.15
tokenizers==0.21.1

# PyTorch (CPU uchun)
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1

# Video / Image processing
opencv-python==4.11.0.86
pillow==11.2.1
numpy<2.0.0

# Database & Auth (agar ishlatsang)
sqlalchemy
alembic
psycopg2-binary
passlib[bcrypt]
python-dotenv
python-jose
pydantic-settings
ffmpeg-python

# JWT uchun
python-jose
passlib[bcrypt]
